reorganize plugin management code
